+++ src/django_mysql/models/fields/__init__.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from django_mysql.models.fields.bit import Bit1BooleanField, NullBit1BooleanField
 from django_mysql.models.fields.dynamic import DynamicField
 from django_mysql.models.fields.enum import EnumField
+from django_mysql.models.fields.fixedchar import FixedCharField
 from django_mysql.models.fields.lists import ListCharField, ListTextField
 from django_mysql.models.fields.sets import SetCharField, SetTextField
 from django_mysql.models.fields.sizes import SizedBinaryField, SizedTextField
@@ -11,6 +12,7 @@ __all__ = [
     "Bit1BooleanField",
     "DynamicField",
     "EnumField",
+    "FixedCharField",
     "ListCharField",
     "ListTextField",
     "NullBit1BooleanField",